run unit tests faster
