+++ package.json
@@ -7,8 +7,8 @@
     "clean": "rimraf out-tsc dist/*",
     "prebuild": "npm run clean",
     "build": "ng-packagr -p ./src/ng-package.json",
-    "build:demo": "rimraf dist && webpack --config ./scripts/webpack.config.js --progress --profile --bail",
-    "start": "webpack-dev-server --config ./scripts/webpack.config.js --inline --progress --port 8080",
+    "build:demo": "rimraf dist && webpack --config ./scripts/webpack.dev.js --progress --profile --bail",
+    "start": "webpack-dev-server --config ./scripts/webpack.dev.js --inline --progress --port 8080",
     "test:watch": "karma start ./scripts/karma.conf.js --no-single-run --auto-watch",
     "test": "karma start ./scripts/karma.conf.js --single-run",
     "coveralls": "cat ./coverage/lcov/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
