+++ tsconfig.json
@@ -15,5 +15,11 @@
       "es2015",
       "dom"
     ]
-  }
+  },
+  "exclude": [
+    "node_modules",
+    "coverage",
+    "integration",
+    "dist"
+  ]
 }
